新建轮播图 以及 分类 还有推荐歌单 并且解决需要登录 301的问题

+++ app/src/main/java/com/example/materialdesignmusic/JSONDATA/BannerData.java
@@ -40,7 +40,7 @@ public class BannerData {
      */
 
     private String pic;
-    private int targetId;
+    private long targetId;
     private Object adid;
     private int targetType;
     private String titleColor;
@@ -80,11 +80,11 @@ public class BannerData {
         this.pic = pic;
     }
 
-    public int getTargetId() {
+    public long getTargetId() {
         return targetId;
     }
 
-    public void setTargetId(int targetId) {
+    public void setTargetId(long targetId) {
         this.targetId = targetId;
     }
